+++ portal/apps/adzone/form.py
@@ -1,6 +1,4 @@
 # -*- coding: utf-8 -*-
-from __future__ import unicode_literals
-
 from django import forms
 
 from .models import BannerAd
@@ -12,6 +10,7 @@ MAX_UPLOAD_SIZE = "150165"  # 150kb
 class UploadFileForm(forms.ModelForm):
 
     def clean(self):
+        cleaned_data = super().clean()
         self.check_file(self)
         return self.cleaned_data
 
@@ -25,23 +24,23 @@ class UploadFileForm(forms.ModelForm):
         if mobile_content:
             mobile_content_excedeed = mobile_content.size > int(MAX_UPLOAD_SIZE)
         if content_excedeed and mobile_content_excedeed:
-            msg_details = u"'Banner escritorio y el Banner móvil' tienen cada uno "
-            msg = u"El " + msg_details + u" un tamaño mayor al máximo permitido (150kb). " + \
-                  u"Por favor, utiliza un banner de un tamaño menor a 150kb."
+            msg_details = "'Banner escritorio y el Banner móvil' tienen cada uno "
+            msg = "El " + msg_details + " un tamaño mayor al máximo permitido (150kb). " + \
+                  "Por favor, utiliza un banner de un tamaño menor a 150kb."
             form._errors["content"] = self.error_class([msg])
             form._errors["mobile_content"] = self.error_class([msg])
             del self.cleaned_data["content"]
             del self.cleaned_data["mobile_content"]
         elif content_excedeed:
-            msg_details = u"'Banner escritorio' tiene"
-            msg = u"El " + msg_details + u" un tamaño mayor al máximo permitido (150kb). " + \
-                  u"Por favor, utiliza un banner de un tamaño menor a 150kb."
+            msg_details = "'Banner escritorio' tiene"
+            msg = "El " + msg_details + " un tamaño mayor al máximo permitido (150kb). " + \
+                  "Por favor, utiliza un banner de un tamaño menor a 150kb."
             form._errors["content"] = self.error_class([msg])
             del self.cleaned_data["content"]
         elif mobile_content_excedeed:
-            msg_details = u"'Banner móvil' tiene"
-            msg = u"El " + msg_details + u" un tamaño mayor al máximo permitido (150kb). " + \
-                  u"Por favor, utiliza un banner de un tamaño menor a 150kb."
+            msg_details = "'Banner móvil' tiene"
+            msg = "El " + msg_details + " un tamaño mayor al máximo permitido (150kb). " + \
+                  "Por favor, utiliza un banner de un tamaño menor a 150kb."
             form._errors["mobile_content"] = self.error_class([msg])
             del self.cleaned_data["mobile_content"]